% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bamm}
\alias{bamm}
\title{bamm}
\usage{
bamm(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run bamm.
}
\details{
Usage: \code{bamm -c <control-file> [--<parameter-name>
<parameter-value> ...]}
}
\examples{
library(outsider)
bamm <- module_import('bamm', repo = 'dombennett/om..bamm')
bamm('--version')
}
